requirements.txt
streamlit>=1.40.0
pandas>=2.0.0
plotly>=5.18.0
fpdf2>=2.7.0
pytest>=7.4.0
sqlalchemy>=2.0.0
alembic>=1.13.0
streamlit-authenticator>=0.3.3
openpyxl>=3.1.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
black>=24.0.0
flake8>=7.0.0
